pyproject.toml
[project]
name = "vizlook"
version = "0.0.0"
description = "The official Python library for the Vizlook API"
authors = [{ name = "Vizlook", email = "support@vizlook.com" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)", 
    "pydantic (>=2.11.7,<3.0.0)"
]
classifiers = [
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Operating System :: POSIX",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux"
]

[project.urls]
Homepage = "https://github.com/vizlook/vizlook-python"
Repository = "https://github.com/vizlook/vizlook-python"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"

